add interaction functionality
