+++ modules/task.js
@@ -18,25 +18,6 @@ export default class Task {
     this.retrieveData();
   }
 
-  updateTask = (text, check) => {
-    const rd = this.taskList;
-    rd.forEach((element) => {
-      if (check === true || check === false) {
-        if (rd[rd.indexOf(element)].title === text.trim()) {
-          rd[rd.indexOf(element)].title = text.trim();
-          rd[rd.indexOf(element)].completed = check;
-        }
-      } else if (check >= 0) {
-        if (rd.indexOf(element) === check) {
-          rd[rd.indexOf(element)].title = text.trim();
-        }
-      }
-    });
-    const taskssString = JSON.stringify(rd);
-    localStorage.setItem('tasksInfo', taskssString);
-    window.location.reload();
-  }
-
   removeItem = (index) => {
     const data = this.taskList;
     data.splice(index, 1);
